requirements.txt
openai==0.27.8
pyaudio==0.2.13
python-dotenv==1.0.0
